Faster approach to code for reusing fragments in fetches
This commit somehwat "inverse" the way fragment reuse is tried.
Before this commit, fragment reuse was tried depth-first, meaning
that we tried reusing fragment on the leaf of the result set and
then up on every selection. Unfortunately, doing so creates a
number of subtlety that force the checks done on every selection
to be somewhat costly, which adds up.

This commit change the process a bit to test for fragment reuse
"at the top level" first (and so always fully expanded selections),
and recurse down for the part that didn't matched anything. In practice,
doing so simplify things a bit due to always dealing with expanded
selections, and this allow to the checks simpler and more
efficient.

Overall, this commit usually improves the time spend for trying fragment
reuse during query planning (sometimes substantially), hence lowering
query planning time.